Rename createPncDisposalFromResult to createPncDisposalsFromResult
