+++ packages/core/phase2/exceptions/HO200205.ts
@@ -4,7 +4,7 @@ import type Exception from "../../types/Exception"
 import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 import checkResultsMatchingPncDisposalsExceptions from "./checkResultsMatchingPncDisposalsExceptions"
 import errorPaths from "../../lib/exceptions/errorPaths"
-import isAmountSpecifiedInResultValid from "../lib/createPncDisposalFromResult/isAmountSpecifiedInResultValid"
+import isAmountSpecifiedInResultValid from "../lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid"
 
 const firstAmountIndex = 0
 const thirdAmountIndex = 2
