--- packages/core/phase2/lib/createPncDisposalFromResult/index.ts
@@ -1,1 +1,0 @@
-export { default as createPncDisposalFromResult } from "./createPncDisposalFromResult"
