+++ packages/core/phase2/lib/createPncDisposalsFromResult/createPncDisposal.ts
