+++ packages/core/phase2/lib/createPncDisposalsFromResult/createPncDisposalsFromResult.ts
@@ -3,11 +3,11 @@ import type { PncDisposal } from "../../../types/PncQueryResult"
 import createPncDisposalByFirstAndSecondDurations from "./createPncDisposalByFirstAndSecondDurations"
 import createPncDisposalByThirdDuration from "./createPncDisposalByThirdDuration"
 
-const createPncDisposalFromResult = (result: Result): PncDisposal[] => {
+const createPncDisposalsFromResult = (result: Result): PncDisposal[] => {
   const pncDisposalByFirstAndSecondDurations = createPncDisposalByFirstAndSecondDurations(result)
   const pncDisposalByThirdDuration = createPncDisposalByThirdDuration(result, pncDisposalByFirstAndSecondDurations.text)
 
   return [pncDisposalByFirstAndSecondDurations, ...(pncDisposalByThirdDuration ? [pncDisposalByThirdDuration] : [])]
 }
 
-export default createPncDisposalFromResult
+export default createPncDisposalsFromResult
