+++ packages/core/phase2/lib/createPncDisposalsFromResult/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.test.ts
