+++ packages/core/phase2/lib/createPncDisposalsFromResult/getDisposalTextFromResult/getDisposalTextFromResult.ts
