+++ packages/core/phase2/lib/createPncDisposalsFromResult/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
