+++ packages/core/phase2/lib/createPncDisposalsFromResult/getDisposalTextFromResult/penaltyPointsDisposalText.ts
