+++ packages/core/phase2/lib/createPncDisposalsFromResult/getDisposalTextFromResult/untilFurtherOrderDisposalText.ts
