+++ packages/core/phase2/lib/createPncDisposalsFromResult/getFirstDateSpecifiedInResult.ts
