+++ packages/core/phase2/lib/createPncDisposalsFromResult/index.ts
@@ -1,0 +1,1 @@
+export { default as createPncDisposalsFromResult } from "./createPncDisposalsFromResult"
