+++ packages/core/phase2/lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid.ts
